+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Add support for using `Config`, `getProject()`, `getStack()`, and `isDryRun()` from Policy Packs
   via upgraded dependency on `@pulumi/pulumi` v1.8.0 (requires v1.8.0 of the Pulumi SDK).
 
+- Provide easier type checking for `validateStack` (https://github.com/pulumi/pulumi-policy/pull/173).
+
 ## 0.3.0 (2019-11-26)
 
 ### Improvements